+++ app/controllers/api/v2/account_lists/donor_accounts_controller.rb
@@ -31,7 +31,7 @@ class Api::V2::AccountLists::DonorAccountsController < Api::V2Controller
   end
 
   def load_donor_account
-    @donor_account ||= DonorAccount.find_by!(uuid: params[:id])
+    @donor_account ||= DonorAccount.find_by_uuid_or_raise!(params[:id])
   end
 
   def render_donor_account
@@ -47,7 +47,7 @@ class Api::V2::AccountLists::DonorAccountsController < Api::V2Controller
   end
 
   def load_account_list
-    @account_list ||= AccountList.find_by!(uuid: params[:account_list_id])
+    @account_list ||= AccountList.find_by_uuid_or_raise!(params[:account_list_id])
   end
 
   def permitted_filters