+++ app/controllers/api/v2/account_lists/notification_preferences_controller.rb
@@ -55,7 +55,7 @@ class Api::V2::AccountLists::NotificationPreferencesController < Api::V2Controll
   end
 
   def load_notification_preference
-    @notification_preference ||= NotificationPreference.find_by!(uuid: params[:id])
+    @notification_preference ||= NotificationPreference.find_by_uuid_or_raise!(params[:id])
   end
 
   def load_notification_preferences
@@ -88,7 +88,7 @@ class Api::V2::AccountLists::NotificationPreferencesController < Api::V2Controll
   end
 
   def load_account_list
-    @account_list ||= AccountList.find_by!(uuid: params[:account_list_id])
+    @account_list ||= AccountList.find_by_uuid_or_raise!(params[:account_list_id])
   end
 
   def pundit_user