+++ app/controllers/api/v2/account_lists/users_controller.rb
@@ -32,7 +32,7 @@ class Api::V2::AccountLists::UsersController < Api::V2Controller
   end
 
   def load_user
-    @user ||= User.find_by!(uuid: params[:id])
+    @user ||= User.find_by_uuid_or_raise!(params[:id])
   end
 
   def render_user
@@ -51,7 +51,7 @@ class Api::V2::AccountLists::UsersController < Api::V2Controller
   end
 
   def load_account_list
-    @account_list ||= AccountList.find_by!(uuid: params[:account_list_id])
+    @account_list ||= AccountList.find_by_uuid_or_raise!(params[:account_list_id])
   end
 
   def pundit_user