+++ app/controllers/api/v2/user/google_accounts_controller.rb
@@ -41,7 +41,7 @@ class Api::V2::User::GoogleAccountsController < Api::V2Controller
   end
 
   def load_google_account
-    @google_account ||= Person::GoogleAccount.find_by!(uuid: params[:id])
+    @google_account ||= Person::GoogleAccount.find_by_uuid_or_raise!(params[:id])
   end
 
   def render_google_account